feat: Use tab in SettingPopup #207
